Update patch set 1
Patch Set 1: Code-Review+1

We moved from RSE to org.eclipse.remote, and the 'chmod +x' might have been a workaround an RSE bug.  So it is quite possible we don't need it anymore.

We know it works on Linux (or at least this is what this patch says).
We don't know about Windows or Mac, but I cannot even confirm if we support debugging with a Windows target or Mac target.  I guess it would be relatively quick for someone with a Windows machine to try doing a remote session to see if it works at all...

Patch-set: 1
Label: Code-Review=+1