Bug 511801 - Remote launch: validate that the remote exec file is absolute
From my experience, bad things happen if the user specifies a
non-absolute path in the box labeled "Remote Absolute File Path for
C/C++ Application".  This patch adds a validation to the tab to make
sure that the path is a valid absolute POSIX path.

This assumes that we do not support remote launching on Windows targets,
and therefore do not need to specify paths such as "C:\foo\bar.exe".

Change-Id: I20367078ff20179f0515272afee17d0986940309
Signed-off-by: Simon Marchi <simon.marchi@polymtl.ca>