Bug 509577 - [Traditional Rendering] Local variables enclosing markings may be wrong in Little Endian presentation
The enclosing markings are meant to reflect contiguous regions of
memory,
however the current Little Endian representation in the traditional
render modifies the memory map to attempt to reflect how the memory
would
be loaded into registers with a size matching the cell size.

The markings are therefore unsuitable for for this special case,
this change removes these markings when the Little Endian formatting is
selected.


Change-Id: I1d7879f8b23329b33e2b17e8c8444d1266b9ef0a