Update patch set 2
Patch Set 2: Code-Review+2 Verified+1

Looks good to me. I have also built the standalone and run it up. It looks like it is working fine. 

(BTW on Oxygen a -clean is needed on each run, see Bug 511961, but that does not have to do with this change.)

Patch-set: 2
Label: Verified=+1
Label: Code-Review=+2