Add printouts to debug unit tests
Currently, the debug unit tests only print which GDB version is requested
and only if tracing is enabled; we have run into bugs where we would
have greatly benefited in easily seeing what version of GDB was
actually launched.

This change makes each test print what version of GDB is requested
and which one is actually run.  It also makes the test print this information
all the time, one line per test, instead of only when tracing is enabled.

Change-Id: Id19d625170b4f956d6205929062c280ac3ecc3b8