Update patch set 2
Patch Set 2: Code-Review+1 Verified+1

(1 comment)

Looks good. Just a minor comment. This will not work with a Windows remote target, which AFAIK is not supported anyway. However it looks like it would work for a Windows host and Linux remote, which is a not so easy but possible case.

Patch-set: 2
Label: Verified=+1
Label: Code-Review=+1