Update patch set 5
Patch Set 5:

I looked through the calls to PDOM.acquireWriteLock() to see if any were not paired with a corresponding releaseWriteLock() in a finally { ... } block. I found one, in PDOMManager.getOrCreatePDOM(). There, we do call releaseWriteLock(), but not in a finally { ... } block, which leaves open the possibility of leaking the lock if the intervening code throws an exception.

I added the finally { ... } block in https://git.eclipse.org/r/91018.

Patch-set: 5