Update patch set 7
Patch Set 7:

> I added the finally { ... } block in https://git.eclipse.org/r/91018.

With that patch on top of this one, both of them timeout.

With this patch on top of that one, that one finishes, and this one times out.

So, this patch itself is somehow causing the timeout? Not sure how, though...

Patch-set: 7
Reviewer: Gerrit User 973 <973@97ee7c02-f12f-4043-b43e-dea463d88b31>
Label: Verified=0