Update patch set 7
Patch Set 7:

> > I added the finally { ... } block in https://git.eclipse.org/r/91018.
 > 
 > With that patch on top of this one, both of them timeout.
 > 
 > With this patch on top of that one, that one finishes, and this one
 > times out.
 > 
 > So, this patch itself is somehow causing the timeout? Not sure how,
 > though...

I did run a jstack at 2:36PM. It was the same pattern in the stacks. Quite odd. I don't know if it would be possible to track the stacks when acquiring/releasing the locks so they they could be printed out? Sometimes I do fStack = new Throwable(); in the contructor of some object and print it out when test fails. Don't know if that strategy could work here.

Patch-set: 7