Update patch set 5
Patch Set 5:

> getting the required id should be immediately computable, without having to query the backend, or even submit stuff to the DSF session executor.

You are right in the sense that the current extensions are able to resolve the Id locally however the comment is to allow other extenders to use other id's which may not be available and may require to query the backend e.g. using the binary name as the Id.

That would make the API future proof and would allow you to provide a fixed value or query the backend for it.

Patch-set: 5