Improve all-stop when mi-async is enabled.
The DSF services and other components were making the assumptions that
in all-stop mode, GDB was not available to receive commands.  With
recent changes, when running GDB 7.12, we run mi-async (i.e.
target-async), even in all-stop.

With this patch, instead of setting the caches to 'target-unavailable'
when the inferior is running in all-stop, we instead only set them
unavailable based on IMIRunControl.isTargetAcceptingCommands().

This allows to show the list of threads in the debug view even when the
target is running, and use e.g., the OS Resource view.

Change-Id: If8be953066d66a90d36f28f94ebaba3871388480