Update patch set 1
Patch Set 1:

Would adding a getEnvVarNormalized() to the EnvironmentReader be an option? It feels to me that a normalization on the CommandLauncher level would be redundant since we already have a normalized map.
Changing the CommandLauncher's getEnvironment to return a normalized map will also mean anyone calling it is now normalized, which might end up causing the same issue you describe.

Another solution would be to create a getEnvironmentVariable private method on CommandLauncher that queries fEnvironment if it is not null and otherwise redirects to the EnvironmentReader.getEnvVar(String) method which in turn is normalized.

Patch-set: 1
Reviewer: Gerrit User 13314 <13314@97ee7c02-f12f-4043-b43e-dea463d88b31>
Label: Verified=0