Update patch set 1
Patch Set 1:

> Would adding a getEnvVarNormalized() to the EnvironmentReader be an
 > option?

It may be a part of the solution. 

 > It feels to me that a normalization on the CommandLauncher
 > level would be redundant since we already have a normalized map.

You have a normalized map in the EnvironmentReader, but the environment that is used in CommandLauncher is derived from EnvironmentReader and then all the user's changes are made to it before the get(PATH) call. 

 > Changing the CommandLauncher's getEnvironment to return a
 > normalized map will also mean anyone calling it is now normalized,
 > which might end up causing the same issue you describe.

I agree.

 > Another solution would be to create a getEnvironmentVariable
 > private method on CommandLauncher that queries fEnvironment if it
 > is not null and otherwise redirects to the EnvironmentReader.getEnvVar(String)
 > method which in turn is normalized.

That works, if the query to fEnvironment is case-insensitive.

============

I wrote the above and now am wondering if I am complicating things too much. Maybe we should just uppercase all the environment variables all the time on Windows as your patch suggests, even the user facing ones.

Patch-set: 1