Generate config.ini for stand-alone debugger based on debug.product
Before this change, we had to remember to edit the config.ini every time
a dependency was modified in debug.product. This change introduces a
script during the build that generates the config.ini with the same list
of plugins specified in debug.product.

The cdtdebug.sh start-up script was simplified in the process because
the config.ini includes all platform-specific plugins now and they just
won't get loaded if they are not a match for the environment.

Change-Id: I658b0dd8688088e203966a154e6154f1b2c851f4
Signed-off-by: Marc-Andre Laperle <marc-andre.laperle@ericsson.com>