Update patch set 6
Patch Set 6:

I would recommend not changing the API and keeping the change to a minimum.
If someone develops a more precise solution that figures out pointers that are typedefs, then we would want to go back to the current API.

So, how about just changing isMofiable() to return true all the time?

Furthermore, this change fixes a typedef pointer not updating, which is the important fix, however it does not allow the user to edit the value of such typedef pointers.   This is because the typedef pointers are still isComplex() which prevents them from being editable.  If we didn't have the optimization of using isComplex() in MIVariableManager.MIVariableObject#getAttributes() then we would ask GDB and properly mark this case as editable.

This makes me wonder if we should instead improve isComplex() to consider any type with a single child as potentially being a pointer and therefore having isComplex() returning false in that case.  That should fix all the issues with typedef pointers and keep some of our optimizations.

What do you think?

Patch-set: 6
Label: Verified=0