Update patch set 6
Patch Set 6:

Hmmm, the problem with changing isComplex as you suggest is that it changes the behavior of the method, so if an extender uses it, we may break something.  If we were able to make a 100% precise implementation of isComplex, then we could pass it as a bug fix, but here we would just switch from having false positives (some non-complex variables reported as complex) to having false negatives (some complex variables not reported as complex).

I think the cleanest thing to do would be to just stop using isComplex in our code and leave isComplex as-is (and maybe document the known problems)?

Patch-set: 6
Label: Verified=0