Make variables defined using a typedef to pointer editable
Since pointers defined using a typedef are wrongfully considered complex
values (see bug 399494), CDT won't let the user manually change their
value.

There are two spots that need fixing:

 - getAttributes, so that the UI will let the user have an editable
   field to enter the new value,
 - writeValue, so that the new value is sent to GDB when the user
   presses enter.

The change in getAttributes means that we will ask GDB whether a
variable is modifiable or not for every variable, without trying to
guess.  It shouldn't introduce many more MI requests, since this is
pretty much only used when the user tries on the field to try to edit
it.  And in the worst case, the value is cached.

The change in writeValue means that we may try to send to -var-update to
GDB for variables that are not modifiable.  In that case, GDB will give
us an error.  But normally, the UI won't let the user edit
non-modifiable values (because of getAttributes), so we won't reach that
point in normal use cases.

Finally, this patch adds a small test to verify that writing a pointer
variable defined using a typedef works correctly.

Change-Id: Idd05fa8b57578bb8177fcf976bdf441c7c0ebab1
Signed-off-by: Simon Marchi <simon.marchi@polymtl.ca>