Update patch set 9
Patch Set 9:

> Hmmm, the problem with changing isComplex as you suggest is that it
 > changes the behavior of the method, so if an extender uses it, we
 > may break something.  
 > If we were able to make a 100% precise
 > implementation of isComplex, then we could pass it as a bug fix,
 > but here we would just switch from having false positives (some
 > non-complex variables reported as complex) to having false
 > negatives (some complex variables not reported as complex).

Isn't the same true for isModifiable()?  We're changing it's behaviour to return false positives.
 
If we look at how we use isComplex() is is better to return false negatives (for which we rely on GDB knowledge instead and get back on our feet) than return false positives, which we don't recover from.

Now, maybe extenders use the method in a different way, but couldn't they use isModifiable() in a way that we could break by changing it?  In then end, we do our best but extenders will need to fix things if we happen to break something.

Patch-set: 9
Reviewer: Gerrit User 457 <457@97ee7c02-f12f-4043-b43e-dea463d88b31>
Label: Verified=0