Don't delete the rcp and rcp-repository directories
When building o.e.cdt.repo we used to delete the
entire directory in which we put the build result.  However, it is
possible that the Standalone debugger was built before o.e.cdt.repo and
already put the valid rcp and rcp-repository directories in that
location.  We therefore should not delete those.

Change-Id: I5d1608bceb48633f2205af8f8282e884ff223351