Update patch set 2
Patch Set 2: Code-Review-2

It works with older GDBs.

However, it does not work in the remote case with GDB/gdbserver 7.12.
The problem is that some commands are rejected by gdbserver, even though we are in mi-async.  And then CDT does not recover properly.  I'll have to look into it.

Patch-set: 2
Label: Code-Review=-2