Update patch set 1
Patch Set 1:

> (1 comment)
 > 
 > This change would cause loading of CDT plugins in workspaces that
 > don't contain any C/C++ projects.

I looked a bit at that. It already does load because of Codan's auto startup, Toggle Breakpoint action, decorators, etc. I was not sure if it really was really a concern given all the other reason why it loads already. But I can check for an alternative.

Patch-set: 1
Label: Verified=0