Update patch set 1
Patch Set 1: Code-Review-1

(2 comments)

This is a nice improvement.

However there is one new case that is problematic in that it appears to the user that updates are in progress but they never complete.

If my selection moves from a "No debug context" (like the launch itself, or the gdb node) then I select a thread (running or suspended) I get  "..." in the disassembly view. I have to select a frame for the update to continue.

This is what I get in my eclipse trace:

	DisassemblyBackendDsf: switch session [0<<null]. Input context=(gdb[0].proc[27705].threadGroup[i1],gdb[0].proc[27705].OSthread[2]).thread[2]
	DisassemblyBackendDsf: switch session done [id=0;context=(gdb[0].proc[27705].threadGroup[i1],gdb[0].proc[27705].OSthread[2]).thread[2];
				frame=null].
				Input context=(gdb[0].proc[27705].threadGroup[i1],gdb[0].proc[27705].OSthread[2]).thread[2]
	DisassemblyPart.debugContextChanged()
	Calling AbstractDocument.replace(0,0,"...\n")
	Adding position to category <category_disassembly> : AddressRangePosition@301af154[INVALID][0:4]->[0:10000000000000000]
	Adding to invalid range list: AddressRangePosition@301af154[INVALID][0:4]->[0:10000000000000000]
	gotoFrame 0 -1
	retrieveFrameAddress 0

In the current master code, if I move from No debug context to running of suspended thread I continue to get "No debug context" message, and if I move from frame to suspended thread I get disassembly still.

Patch-set: 1
Label: Code-Review=-1