Update patch set 2
Patch Set 2: Code-Review-2

(1 comment)

With patch set #2 everytime I click on an frame the view is refreshing from the backend. This didn't happen in patch set #1. 

The result is that clicking and down the stack is now laggy as there is constant calls to GDB.

I have marked the suspicious line. 

(PS If you are joining the CDT call in a minute we can discuss further if you want.)

Patch-set: 2
Label: Code-Review=-2