Bug 515214 - Fake Include Guard PoC to improve performance.
In case no include guard was found in a file, generate a fake include
guard based on the filename in order to improve performance for headers
included in LOTS of different contexts.
TODO This is a proof-of-concept only. This code should be turned on
only conditionally, with a UI preference or System Property.

Change-Id: I4fe91f0015bcd042e5b156e04b102a284f37d6af
Signed-off-by: Martin Oberhuber <martin.oberhuber@windriver.com>