Update patch set 3
Patch Set 3:

> I don't think cdt_9_2 should get new features as it is a
 > maintenance branch.

This commit was not intended to be pushed to the 9.2 branch. I pushed it to the wrong branch as I rebased it to 9.2 for testing it with our plug-ins that depend on const placement. Sorry for the confusion.

Patch-set: 3
Reviewer: Gerrit User 661 <661@97ee7c02-f12f-4043-b43e-dea463d88b31>
Label: Verified=0