Update patch set 1
Patch Set 1: Code-Review-2

(8 comments)

I have reviewed this, there are a number of changes that need to be made that I have tried to outline.

In addition to these comments, as I don't have visibility on any code that actually has an implementation of ICPropertyTab.isValid() I am a bit unsure on one part that I will try to explain.

The normal way that validity is handled is to call #setValid() so that the #isValid calls have little to do. That may be what you are doing, but I can't tell if the ICPropertyTab should therefore be "pushing" the setValid up to the AbstractPage instead of it being pulled. 

I don't object to the pull, and I leave it up to your implementations to handle that properly, but did wonder if you have considered this and if the current implementation works with that model.

Patch-set: 1
Reviewer: Gerrit User 27769 <27769@97ee7c02-f12f-4043-b43e-dea463d88b31>
Label: Code-Review=-2