Update patch set 5
Patch Set 5:

It works well for the local case but I'm not sure what to expect for the remote case. If I run with gdbserver 7.12, it doesn't seem to update the list of threads in the debug view while the target is running. My program creates new threads in a while loop with a sleep (a slow thread bomb), I can see the threads being added as it runs in the Debug view when using local gdb but not in the case of remote. So I guess my question is, what is expected to have changed with this patch for the remote case?

Patch-set: 5
Reviewer: Gerrit User 598 <598@97ee7c02-f12f-4043-b43e-dea463d88b31>
Label: Verified=0