Update patch set 1
Patch Set 1:

> See comments on the 9.2 version of this change.

Thanks, I'm copying them here now.

 > This isn't correct. The descriptor enablement should match what's in 
 > CoreBuildLaunchDescriptorType.getDescriptor() 
 > and make sure that the project also supports the Core Build system. 
 > You'll probably need to add a propertyTester and add it as a third clause 
 > (after the nature check).
 > Also the launch config provider enablement is wrong. 

You probably mean it is *not complete*? Wrong is the current state, because current code has NO enablement at all and so it is always enabled, if looking into LaunchDescriptorTypeInfo and LaunchConfigProviderInfo

 > The enablement gets passed a launch descriptor and a launch target 
 > and the enablement must pass for both.

So the first check for descriptor is same as in CoreBuildLaunchDescriptorType.getDescriptor(), but where is the second for target?

Patch-set: 1
Reviewer: Gerrit User 1179 <1179@97ee7c02-f12f-4043-b43e-dea463d88b31>
Label: Verified=0