Update patch set 1
Patch Set 1:

> You probably mean it is *not complete*? Wrong is the current state,
 > because current code has NO enablement at all and so it is always
 > enabled, if looking into LaunchDescriptorTypeInfo and
 > LaunchConfigProviderInfo

Well, the descriptor is more complete than it was, yes, but the config provider will actually always fail. It never gets fed an IProject.

My point really is that we didn't implement this yet because it's going to be a fair amount of work to get right and we didn't have the time. Thanks for helping us get there!

 > So the first check for descriptor is same as in CoreBuildLaunchDescriptorType.getDescriptor(),
 > but where is the second for target?

It's here: CoreBuildLocalLaunchConfigProvider.supports(). It basically checks if the target is the Local one. For the descriptor, there should be an adapter to IProject so you could use a similar check.

BTW, it would be awesome to have some JUnits for this as well to make sure the enablement is working correctly.

Patch-set: 1