Update patch set 1
Patch Set 1:

> I think this should use the plugin version, not the feature.

The primary motivation for this feature, as described in the bug report, is to allow users to work around deficiencies / unsupported features in CDT by conditionally including some code that is only processed by some versions of CDT.

Given this use case, I think it makes more sense to use the user-facing version number, which is the feature version. This is what shows up in the "Target Milestone" field in bugzilla (if e.g. the user is working around a specific bug), and in our public messaging on pages such as https://www.eclipse.org/cdt/. The individual version numbers of plugins are something users may not even be aware of, and they may not know how to map CDT versions to plugin versions.

 > Plugins are not necessarily in features. Anyone can put the
 > org.eclipse.cdt.core plugin in another feature or build a product
 > based on plugins (the stand-alone debugger does that).

Products that include the org.eclipse.cdt.core plugin without the whole org.eclipse.cdt feature will fall back to the current behaviour of defining __CDT_PARSER__ as 1.

Patch-set: 1