Bug 516836 Terminal support for Arduino.
Add Arduino Serial Monitor as a Terminal Connector type. It uses the
Terminal's new cdtserial connector for communication and adds a
config panel that lets you pick the Arduino remote connection to
auto select the serial port. Also switched the Arduino launch
to fetch the serial port out of the global registry to pause and
resume the port while launching. The Terminal does not use the
remote connection to get the serial port.

Note: also adding the terminal as a dependency for builds. Yes,
further adds to our build cycle, but we're dealing with it so far
with o.e.remote.

Change-Id: I5a7cc9ffbf8ae60cd86e0e5440547d12202acbac