A basic LSP4E integration with Clangd
This plug-in adds an simple integration between LSP4E and Clangd.
It simply launches clangd when opening C/C++ files with the
Generic Editor.

To try it, you need to:
- Import the projects in your Eclipse workspace (you can only import the
lsp4-cpp ones, the sources for CDT are not needed)
- Open lsp4e.cpp-staging.target and set it as the target platform
- You need to have clangd built and on the PATH. See
https://clang.llvm.org/get_started.html
- Start Eclipse: right-click on a project > Run as > Eclipse Application
- In the new Eclipse instance, you can right-click on a C/C++ file >
Open As > Generic Editor. This should make use of clangd, providing some
basic diagnostics and formatting commands.

This should help the Eclipse community to start playing with Clangd.

Change-Id: I688d3a952dbc80d9f61186178f3c8a033afcc587
Signed-off-by: Marc-Andre Laperle <marc-andre.laperle@ericsson.com>