Update patch set 6
Patch Set 6:

I had a slightly different change in mind for lookInside(), but instead of trying to describe it in words I just made it and pushed an updated patch.

Michi, could you verify that this updated patch works for your use cases (particularly the ones you don't have reduced testcases for)?

Patch-set: 6