Update patch set 1
Patch Set 1:

(4 comments)

Open Declaration has a test suite, in org.eclipse.cdt.ui.tests.selection, so I would like to have some tests for the change to OpenDeclarationsJob. CPPSelectionTestsNoIndexer.java is a good place to put them.

It would be a good to have a test for each of the simple-declaration, parameter-declaration, and type-id cases.

(There is no need to have separate tests for 'typeof'.)

Patch-set: 1
Label: Verified=0