Update patch set 5
Patch Set 5:

(2 comments)

Thanks for the example it helps me understand the underlying problem. I don't see how this patchset fixes it though.

Are you looking for more advice on how to proceed? If so I'll have to investigate, but as I said below you may be the most experienced in this area.

-- below are my responses to the line comments --

Patch-set: 5