Update patch set 5
Patch Set 5:

What happens is a bit quick and dirty (I feel like rewriting this class).
After usedOutType is set to null there is a call to addRuleForSource which calls which calls calculateOutputsForSource which sets usedOutType under certain conditions.
...
And explaining I think..; this is not a cdt solution but a quick and dirty solution to get it to work for me based on language provier usage :-( :-(.
I need to think a bit more about this but maybe it is just like in my comment and it should be a loop over tool.getPrimaryOutputTypes()

Patch-set: 5