Update patch set 1
Patch Set 1:

> can someone help me finding out how to load an external header file
 > in the codeassist unit tests? thx

Good question. There is a way but it's not very obvious / well documented.

All completion test cases include a common header file, whose contents are found in a comment at the bottom of CompletionTestBase.java.

So, you can add the header code to that comment (try to re-use existing parts of the code if possible), and then the .cpp code as usual in the comment above the testcase.

Patch-set: 1
Label: Verified=0