Update patch set 4
Patch Set 4:

Hi Norman, I am reviewing this now. This is showing up as a huge (+1827/-1397) change that would require a CQ. This is due to the EOL change on MemoryBrowser.java. Can you resubmit with the same newlines as are already in the file. If the current EOL in master is not LF, please let me know and I will correct that as a separate commit.

Patch-set: 4