Update patch set 2
Patch Set 2:

> there are soooo many way to mess things up ;-)
 > 
 > Test<Model<1300>::family_t::value>::/*cursor*/

Thanks. I can reproduce the issue now.

The remaining problem was that even though we deferred the execution of the function call to after instantiation, we weren't marking the evaluation as dependent. This resulted in constant<getFamily()> being a CPPClassInstance rather than a CPPDeferredClassInstance, and so instantiateType() didn't know that the argument still needed to be instantiated.

This is addressed by a small change to the patch.

Patch-set: 2