Update patch set 2
Patch Set 2:

(3 comments)

Michi, thanks for giving this a try! This is something I've been wanting to fix for a long time. (Apart from this bug, this change will contribute to making the original example in bug 472818 work as well.)

I haven't fully thought through the implications of calling CPPTemplates.selectSpecialization() with dependent arguments, but HeuristicResolver seems like a safe enough place to experiment with things like this, so I say we go for it!

Patch-set: 2