Update patch set 3
Patch Set 3:

> Nathan, I added the test case from the other bug and fixed
 > HeuristicResolver accordingly.
 > However, one thing remains that I can't wrap my head around just
 > yet. When you use the -> operator on the returned iterator I don't
 > know where I should do the dereferencing; that would get us the
 > template parameter for which we can't add proposals of course.
 > Right now we also get the iterator bindings.

The dereferencing is handled in EvalMemberAccess.getFieldOwnerType().

Notice the function has an early exit if "scope instanceof ICPPInternalUnknownScope", which is what happens in this case.

We'd somehow have to propagate an "isPrefixLookup" flag there, and if so, get into CPPSemantics.findOverloadedOperator(), and have it call setPrefixLookup(true) on the LookupData it creates. (That will ensure we get into the "lookup.isPrefixLookup()" branch of CPPUnknownTypeScope.getBindings(), which will use HeuristicResolver to get us useful results from the unknown type scope.)

Patch-set: 3