Update patch set 4
Patch Set 4:

On a general note. I have the impression that the changes introduce quite some lag into the content assist. This is just an impression and could as well have to do with my workspace (has quite some old cruft) and that I'm running the IDE in the debugger all the time. I just wanted to mention it.

Patch-set: 4