Update patch set 4
Patch Set 4:

> On a general note. I have the impression that the changes introduce
 > quite some lag into the content assist. This is just an impression
 > and could as well have to do with my workspace (has quite some old
 > cruft) and that I'm running the IDE in the debugger all the time. I
 > just wanted to mention it.

Interesting. Perhaps selectSpecialization() is a fairly heavy-weight operation. We could try to optimize it a bit. For example, I don't think the call to instantiatePartialSpecialization() at the end is necessary for the heuristic use case - we really just want the template.

We could commit the patch on the master branch, use it for a while in our real-world projects, and see if it causes a noticeable slowdown? The next release is 9.4 in December, so we can revert it if we notice a slowdown before then.

The other thing we could do is use a tool like VisualVM to get before-and-after profiles of a content assist operation and compare them.

Patch-set: 4