Update patch set 1
Patch Set 1: Code-Review-1

This is not the right solution approach. A named-type-specifier should only offer completions that could form (the beginning of) a valid named-type-specifier. Notice how all the existing bindings listed in that 'if' condition are types.

A non-type template parameter, however, isn't a type; it's something that would appear in an expression, not a type specifier.

The way this is supposed to work is:

  - The parser builds an ambiguous template argument node,
    which contains variants for both type and non-type
    template argument.

  - Both variants contain a name, and both names are
    added to the IASTCompletionNode.

  - In DOMCompletionProposalComputer, the
    "for (IASTName name : names)" loop will therefore
    iterate twice.

  - In one iteration, the name will be inside a 
    named-type-specifier, and we collect completions
    that would form a valid type specifier.

  - In the other iteration, the name will be inside
    an id-expression. CPPASTIdExpression.findBindings()
    will be called, and we get the completions that
    would form a valid expression, including the one
    for the non-type template parameter.

However, it looks like we actually get just one name instead of two. We should investigate why the parser is not adding the second name to the completion node.

Patch-set: 1
Reviewer: Gerrit User 973 <973@97ee7c02-f12f-4043-b43e-dea463d88b31>
Label: Code-Review=-1